pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tta"
version = "0.1.0"
description = "Therapeutic Text Adventure — AI-powered narrative game"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.135",
    "uvicorn[standard]>=0.30",
    "sqlmodel>=0.0.38",
    "asyncpg>=0.30",
    "alembic>=1.13",
    "litellm>=1.83,<2",
    "neo4j>=5.0",
    "redis>=5.0",
    "structlog>=24.0",
    "langfuse>=4.0,<5.0",
    "pydantic-settings>=2.0",
    "jinja2>=3.1",
    "tenacity>=9.0",
    "httpx>=0.27",
    "prometheus-client>=0.24.1",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.40.0",
    "pyjwt>=2.9",
    "bcrypt>=4.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-bdd>=7.0",
    "pytest-timeout>=2.1",
    "hypothesis>=6.0",
    "pytest-testmon>=2.0",
    "pytest-watcher>=0.4",
    "ruff>=0.8",
    "pyright>=1.1",
]
audit = ["pip-audit"]

[tool.hatch.build.targets.wheel]
packages = ["src/tta"]

# ---------------------------------------------------------------------------
# Ruff — linting + formatting
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # Formatter conflicts
    "COM812",   # missing trailing comma (ruff format handles this)
    "ISC001",   # implicit string concat (ruff format handles this)

    # Docstrings — enable progressively as docs are added
    "D",

    # Type annotation strictness — pyright at standard mode handles this
    "ANN",

    # FastAPI / SQLModel idioms
    "B008",     # Depends() in function defaults — idiomatic FastAPI
    "RUF012",   # mutable ClassVar — SQLModel field() pattern

    # Exception message style
    "EM101",    # exception msg from string literal
    "EM102",    # exception msg from f-string
    "TRY003",   # long messages outside exception class

    # Boolean flags — common in settings / config classes
    "FBT001",   # boolean positional arg
    "FBT002",   # boolean default arg
    "FBT003",   # boolean positional value in call

    # TODO / FIXME tracking — handled in GitHub issues, not inline
    "FIX002",
    "TD002",
    "TD003",
    "ERA001",   # commented-out code — sometimes intentional during dev

    # Conditional imports — common with TYPE_CHECKING and lazy loading
    "PLC0415",

    # Unused arguments — FastAPI dependency injection creates many stubs
    "ARG001",   # unused function argument
    "ARG002",   # unused method argument

    # Blind exception catching — intentional in resilience / error boundary code
    "BLE001",

    # Complexity metrics — tech debt to address separately, not blocking
    "C901",     # too complex
    "PLR0912",  # too many branches
    "PLR0913",  # too many function arguments
    "PLR0915",  # too many statements

    # Logging — structlog uses structured calls, not %-format
    "G201",     # logging-warn (structlog API differs)
    "G004",     # logging f-string — fine with structlog bind()

    # Possible hardcoded password — false positives on "Bearer", "PASS", "TOKEN" constants
    "S105",

    # Ambiguous unicode — en-dashes and similar are intentional in strings/comments
    "RUF002",
    "RUF003",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",     # assert — standard pytest pattern
    "S105",     # hardcoded password — test fixture defaults
    "S106",     # hardcoded password in argument — test fixture defaults
    "PLR2004",  # magic values — OK in test assertions
    "SLF001",   # private member access — OK in unit tests
    "PT019",    # fixture without return injected as param — _setup-style fixtures
]
"migrations/**/*.py" = ["ALL"]
"scripts/**/*.py" = ["T201", "INP001", "PLR2004", "S101"]
"plans/**/*.py" = ["ALL"]
"specs/**/*.py" = ["ALL"]
"prompts/**/*.py" = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["tta"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ---------------------------------------------------------------------------
# Pyright — type checking
# ---------------------------------------------------------------------------
[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
include = ["src"]
venvPath = "."
venv = ".venv"

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: Unit tests (no external services)",
    "integration: Integration tests (require Docker services)",
    "e2e: End-to-end tests",
    "slow: Slow tests (>5s)",
    "neo4j: Requires Neo4j",
    "redis: Requires Redis",
    "postgres: Requires PostgreSQL",
    "golden: Golden file comparison tests",
    "flaky: Known flaky tests",
    "llm_live: Tests that call real LLM APIs",
    "hypothesis: Property-based tests",
    "bdd: BDD acceptance tests",
]

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["tta"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@overload",
    "raise NotImplementedError",
]
